Update node version to 8.8.1
Versions 4.8.2 and later, 6.10.2 and later, as well as all versions of 8.x are
vulnerable to an issue that can be used by an external attacker to cause a
denial of service. The severity of this vulnerability is HIGH and users of the
affected version should plan to upgrade when a fix is made available.

https://archive.is/LRy8f